Add files via upload
A Python program was developed to aid in the analysis of magnetic hysteresis curves, extracting useful information from experimental measurements performed in VSM. The code has a simple graphical interface (built in Tkinter) and integrates functionalities for reading, processing, visualizing, and exporting data. This program is useful for researchers and students working on: Magnetic characterization of materials; Separation of ferromagnetic and antiferromagnetic contributions; Rapid analysis of hysteresis curves obtained in experiments.